+++ macros/row_access_policies/add_row_access_policy.sql
@@ -1,4 +1,4 @@
-{% macro add_row_access_policy(materlization, model_schema, model_name, column_name, policy_name, columns) %}
+{% macro add_row_access_policy(materlization, model_schema, model_name, policy_name, columns) %}
     {{ log("Adding row access policy " ~ policy_name ~ " for model " ~ model_schema|lower ~ "." ~ model_alias|lower, info=True) }}
     {%- call statement('set_statement', fetch_result=True) -%}
         alter {{materlization}} {{model_schema|upper}}.{{model_name|upper}} add row access policy {{ var("data_governance_database") }}.{{ var("policy_store") }}.{{ policy_name }} on ( {{columns|join(',')}} );